add CentOS 5.11 to test matrix
This commit adds CentOS 5.11 to the test matrix. This requires working
around a bug in the bento box where the cacert.pem bundle doesn’t have
the VeriSign certificates that AWS uses.

https://github.com/chef/bento/issues/325

This workaround is applied as a recipe in the test cookbook because it
may not be necessary for end users who may not have that certificate
bundle.