Use product keys as the name attribute
Fixes #1.

Implements a product matrix and product_lookup method that returns a
hash of product information based on the table in #1, with naming
conventions updated to upcoming product updates. It attempts to be
smart so users of the chef_ingredient resource don't need to know the
underlying details about our package naming, especially wrt old
package names.

The product_lookup method also validates that the chef-server package
is selected properly based on the available, supported versions (12
and 11), and fails if the version is specified as > 1.0 and < 11.0,
since we did not publish omnibus packages for those versions.

Add the version attribute to the omnibus_service resource so it can
properly look up the ctl command.