Fixes #26, remove the ownership and permissions
Setting the owner and permissions on the config file's directory
conflicts with product ctl commands that also manage those
directories. For example some have the ownership set to opscode
instead of root, meaning this is updated every time and causes a
reconfigure to run again, extending the length of time it takes to
converge a node.