Fix rspec failures and warnings
Add provides in providers to get rid of noisy warnings

Test that we're getting the right provider (yum or apt) for the
different platforms by using ChefSpec's runner lookup method
`#find_resource`. We want to ensure that the returned resource is not
nil, uses the correct resource class, and finally that it's a package
resource with an install action.