Fix permissions on chef-backend-secrets.json
`/etc/chef-backend/chef-backend-secrets.json` must be readable by the user that the `chef-backend` services are running as in-order for cluster actions like leader failover and promotion to work.

This PR changes the mode of the secrets file from `0600` to `0640`, makes the user and group associated with the secrets file configurable, and defaults the file ownership to `root:chef_pgsql` (which is required for cluster operations to work when using default settings).

Two new properties have been added to this resource to allow for the secrets file ownership to be configurable:

```
chef_backend_secrets_user
chef_backend_secrets_group
```

Signed-off-by: Craig Thayer <cthayer@craigthayer.com>