Add rubygems_url parameter to chef_ingredient
This adds a rubygems_url to the chef_ingredient resource which will allow
community distributions to install their patched version of mixlib-install and
thus allow this resource to work as-is otherwise for their users.

This mimics the rubygems_url parameter used in the chef_client_updater cookbook
and resource for similar reasons.

Signed-off-by: Lance Albertson <lance@osuosl.org>